pyproject.toml
[project]
name = "perplexity-api-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.97.1",
    "python-dotenv>=1.0.0",
]
